Split things
