+++ egi_notebooks_hub/egiauthenticator.py
@@ -12,11 +12,10 @@ from oauthenticator.generic import GenericOAuthenticator
 from tornado.httpclient import (
     AsyncHTTPClient,
     HTTPClientError,
-    HTTPError,
     HTTPRequest,
 )
 from tornado.httputil import url_concat
-from traitlets import Bool, List, Unicode, default, validate
+from traitlets import List, Unicode, default, validate
 
 
 class EGICheckinAuthenticator(GenericOAuthenticator):
@@ -144,181 +143,3 @@ class EGICheckinAuthenticator(GenericOAuthenticator):
         auth_state = await user.get_auth_state()
         if auth_state and callable(getattr(user.spawner, "set_access_token", None)):
             user.spawner.set_access_token(auth_state["access_token"])
-
-
-class DataHubAuthenticator(EGICheckinAuthenticator):
-    """
-    EGI Check-in + datahub authenticator for JupyterHub
-    Uses OpenID Connect with aai.egi.eu, fetches DataHub token
-    and keeps it in auth_state
-    """
-
-    onezone_url = Unicode(
-        "https://datahub.egi.eu", config=True, help="""Onedata onezone URL"""
-    )
-
-    oneprovider_host = Unicode(
-        "plg-cyfronet-01.datahub.egi.eu",
-        config=True,
-        help="""Onedata oneprovider hostname""",
-    )
-
-    onepanel_url = Unicode(
-        "",
-        config=True,
-        help="""Endpoint of the oneprovider to establish mappings,
-                if undefined, it will use https://<oneprovider_host>:9443/""",
-    )
-
-    oneprovider_token = Unicode("", config=True, help="""Onedata oneprovider token""")
-
-    map_users = Bool(False, config=True, help="""perform mapping""")
-
-    oneclient_token_name = Unicode(
-        "oneclient.notebooks.egi.eu",
-        config=True,
-        help="""Name of token in the onezone for oneclient""",
-    )
-
-    onezone_token_name = Unicode(
-        "onezone.notebooks.egi.eu",
-        config=True,
-        help="""Name of token in the onezone for the onezone""",
-    )
-
-    storage_id = Unicode("", config=True, help="""Storage id to use for mapping""")
-
-    async def create_onedata_token(self, checkin_token, token_name, caveats):
-        onedata_token = None
-        onedata_user = None
-        http_client = AsyncHTTPClient()
-        headers = {
-            "content-type": "application/json",
-            "x-auth-token": "egi:%s" % checkin_token,
-        }
-        token_url = (
-            self.onezone_url + "/api/v3/onezone/user/tokens/named/name/%s" % token_name
-        )
-        req = HTTPRequest(token_url, headers=headers, method="GET")
-        try:
-            resp = await http_client.fetch(req)
-            datahub_response = json.loads(resp.body.decode("utf8", "replace"))
-            onedata_token = datahub_response["token"]
-            onedata_user = datahub_response["subject"]["id"]
-            self.log.debug("Reusing existing token!")
-        except HTTPError as e:
-            if e.code != 404:
-                self.log.info("Something failed! %s", e)
-                raise e
-        if not onedata_token:
-            # we don't have a token, create one
-            token_desc = {
-                "name": token_name,
-                "type": {"accessToken": {}},
-                "caveats": caveats,
-            }
-            req = HTTPRequest(
-                self.onezone_url + "/api/v3/onezone/user/tokens/named",
-                headers=headers,
-                method="POST",
-                body=json.dumps(token_desc),
-            )
-            try:
-                resp = await http_client.fetch(req)
-                datahub_response = json.loads(resp.body.decode("utf8", "replace"))
-                onedata_token = datahub_response["token"]
-            except HTTPError as e:
-                self.log.info("Something failed! %s", e)
-                raise e
-            # Finally get the user information
-            req = HTTPRequest(
-                self.onezone_url + "/api/v3/onezone/user",
-                headers=headers,
-                method="GET",
-            )
-            try:
-                resp = await http_client.fetch(req)
-                datahub_response = json.loads(resp.body.decode("utf8", "replace"))
-                onedata_user = datahub_response["userId"]
-            except HTTPError as e:
-                self.log.info("Something failed! %s", e)
-                raise e
-        return onedata_token, onedata_user
-
-    async def authenticate(self, handler, data=None):
-        user_data = await super(DataHubAuthenticator, self).authenticate(handler, data)
-        checkin_token = user_data["auth_state"]["access_token"]
-        caveats = [{"type": "interface", "interface": "oneclient"}]
-        oneclient_token, onedata_user = await self.create_onedata_token(
-            checkin_token, self.oneclient_token_name, caveats
-        )
-        caveats = [
-            {"type": "interface", "interface": "rest"},
-            {"whitelist": ["ozw-onezone"], "type": "service"},
-        ]
-        onezone_token, _ = await self.create_onedata_token(
-            checkin_token, self.onezone_token_name, caveats
-        )
-        onedata_info = {
-            "oneclient_token": oneclient_token,
-            "onedata_user": onedata_user,
-            "oneprovider": self.oneprovider_host,
-            "onezone_url": self.onezone_url,
-            "onezone_token": onezone_token,
-        }
-        user_data["auth_state"].update(onedata_info)
-        return user_data
-
-    async def pre_spawn_start(self, user, spawner):
-        await super(DataHubAuthenticator, self).pre_spawn_start(user, spawner)
-        auth_state = await user.get_auth_state()
-        if not auth_state:
-            # auth_state not enabled
-            return
-        if self.map_users:
-            if self.onepanel_url:
-                map_url = self.onepanel_url
-            else:
-                map_url = f"https://{self.oneprovider_host}:9443"
-            map_url += (
-                f"/api/v3/onepanel/provider/storages/{self.storage_id}"
-                "/luma/local_feed/storage_access/all"
-                "/onedata_user_to_credentials"
-            )
-            headers = {
-                "content-type": "application/json",
-                "x-auth-token": self.oneprovider_token,
-            }
-            http_client = AsyncHTTPClient()
-            user_id = auth_state.get("onedata_user")
-            req = HTTPRequest(map_url + f"/{user_id}", headers=headers, method="GET")
-            try:
-                resp = await http_client.fetch(req)
-                self.log.info("Mapping exists: %s", resp.body)
-            except HTTPError as e:
-                if e.code == 404:
-                    mapping = {
-                        "onedataUser": {
-                            "mappingScheme": "onedataUser",
-                            "onedataUserId": user_id,
-                        },
-                        "storageUser": {
-                            "storageCredentials": {"type": "posix", "uid": "1000"},
-                            "displayUid": "1000",
-                        },
-                    }
-                    req = HTTPRequest(
-                        map_url,
-                        headers=headers,
-                        method="POST",
-                        body=json.dumps(mapping),
-                    )
-                    try:
-                        resp = await http_client.fetch(req)
-                        self.log.info("Mapping created: %s", resp.body)
-                    except HTTPError as e:
-                        self.log.info("Something failed! %s", e)
-                        raise e
-                else:
-                    self.log.info("Something failed! %s", e)
-                    raise e
