+++ egi_notebooks_hub/egispawner.py
@@ -2,15 +2,12 @@
 """
 
 import base64
-import json
 import uuid
 
 from kubernetes.client import V1ObjectMeta, V1Secret
 from kubernetes.client.rest import ApiException
 from kubespawner import KubeSpawner
-from tornado import web
-from tornado.httpclient import AsyncHTTPClient, HTTPError, HTTPRequest
-from traitlets import Bool, Dict, List, Unicode
+from traitlets import Unicode
 
 
 class EGISpawner(KubeSpawner):
@@ -137,167 +134,3 @@ class EGISpawner(KubeSpawner):
         if id_token:
             data["id_token"] = base64.b64encode(id_token.encode()).decode()
         self._update_token_secret(data)
-
-
-class DataHubSpawner(EGISpawner):
-    onedata_user_env = Unicode(
-        "ONEDATA_USER",
-        config=True,
-        help="""Environment variable that contains the onedata user""",
-    )
-
-    onezone_env = Unicode(
-        "ONEZONE_URL",
-        config=True,
-        help="""Environment variable that contains the onezone URL""",
-    )
-
-    token_env = Unicode(
-        "ONECLIENT_ACCESS_TOKEN",
-        config=True,
-        help="""Name of the environment variable to store the token""",
-    )
-
-    onezone_token_env = Unicode(
-        "ONEZONE_ACCESS_TOKEN",
-        config=True,
-        help="""Name of the environment variable to store the token for onezone""",
-    )
-
-    oneprovider_env = Unicode(
-        "ONEPROVIDER_HOST",
-        config=True,
-        help="""Name of the environment variable to store the oneprovider
-                host""",
-    )
-
-    force_proxy_io = Bool(False, config=True, help="""Force the use of proxied I/O""")
-
-    force_direct_io = Bool(False, config=True, help="""Force the use of direct I/O""")
-
-    only_local_spaces = Bool(
-        True, config=True, help="""Only mount those spaces local to the provider"""
-    )
-
-    mount_point = Unicode(
-        "/mnt/oneclient",
-        config=True,
-        help="""Mountpoint for oneclient""",
-    )
-
-    sidecar_image = Unicode(
-        "eginotebooks/oneclient-sidecar",
-        config=True,
-        help="""Oneclient image to use""",
-    )
-
-    sidecar_resources = Dict(
-        {
-            "requests": {"memory": "512Mi", "cpu": "250m"},
-            "limits": {"memory": "1Gi", "cpu": "500m"},
-        },
-        config=True,
-        help="""resource limits for the sidecar""",
-    )
-
-    oneprovider_storage_mapping = List(
-        [],
-        config=True,
-        help="""
-        List of dicts like:
-            {"storage_id": "<oneprovider storage id>",
-             "mount_point": "volume mount point"}
-        """,
-    )
-
-    extra_mounts = List([], config=True, help="""extra volume mounts in k8s""")
-
-    def auth_state_hook(self, spawner, auth_state):
-        # get onedata stuff ready to be used later on
-        spawner.environment[self.token_env] = auth_state.get("oneclient_token")
-        spawner.environment[self.onezone_env] = auth_state.get("onezone_url")
-        spawner.environment[self.oneprovider_env] = auth_state.get("oneprovider")
-        spawner.environment[self.onezone_token_env] = auth_state.get("onezone_token")
-        spawner.environment[self.onedata_user_env] = auth_state.get("onedata_user")
-
-    async def _get_local_spaces(self, oneprovider_host, onezone_url, onezone_token):
-        # 1. Get the id of the oneprovider
-        http_client = AsyncHTTPClient()
-        req = HTTPRequest(
-            f"https://{oneprovider_host}/api/v3/oneprovider/configuration", method="GET"
-        )
-        try:
-            resp = await http_client.fetch(req)
-        except HTTPError as e:
-            self.log.warning("Unable to connect to oneprovider: %s", e)
-            raise web.HTTPError(403)
-        resp_json = json.loads(resp.body.decode("utf8", "replace"))
-        provider_id = resp_json.get("providerId", None)
-        if not provider_id:
-            self.log.warning("Unable to get provider id: %s", resp_json)
-            raise web.HTTPError(403)
-        # 2. Get the spaces supported by the oneprovider
-        req = HTTPRequest(
-            f"{onezone_url}/api/v3/onezone/providers/{provider_id}/spaces",
-            method="GET",
-            headers={"X-Auth-Token": f"{onezone_token}"},
-        )
-        try:
-            resp = await http_client.fetch(req)
-        except HTTPError as e:
-            self.log.warning("Unable to get spaces from onezone: %s", e)
-            raise web.HTTPError(403)
-        resp_json = json.loads(resp.body.decode("utf8", "replace"))
-        self.log.debug(resp_json.get("spaces", []))
-        return resp_json.get("spaces", [])
-
-    async def pre_spawn_hook(self, spawner):
-        host = spawner.environment.get(self.oneprovider_env, "")
-        token = spawner.environment.get(self.token_env, "")
-        onezone_url = spawner.environment.get(self.onezone_env, "")
-        onezone_token = spawner.environment.get(self.onezone_token_env, "")
-        cmd = ["oneclient", "-f", "-H", f"{host}"]
-        if self.only_local_spaces:
-            # limit the spaces we mount to avoid issues
-            for space_id in await self._get_local_spaces(
-                host, onezone_url, onezone_token
-            ):
-                cmd.extend(["--space-id", f"{space_id}"])
-            self.log.debug("CMD: %s", cmd)
-        if self.force_proxy_io:
-            cmd.append("--force-proxy-io")
-        if self.force_direct_io:
-            cmd.append("--force-direct-io")
-        if self.oneprovider_storage_mapping:
-            for mapping in self.oneprovider_storage_mapping:
-                cmd.append("--override")
-                cmd.append("%(storage_id)s:mountPoint:%(mount_point)s" % mapping)
-        cmd.append(self.mount_point)
-        volume_mounts = [
-            {"mountPath": f"{self.mount_point}:shared", "name": "oneclient"},
-        ]
-        if self.extra_mounts:
-            volume_mounts.extend(self.extra_mounts)
-        spawner.extra_containers = [
-            {
-                "name": "oneclient",
-                "image": self.sidecar_image,
-                "env": [
-                    {"name": self.oneprovider_env, "value": host},
-                    {"name": self.token_env, "value": token},
-                ],
-                "resources": self.sidecar_resources,
-                "command": cmd,
-                "securityContext": {
-                    "runAsUser": 1000,
-                    "privileged": True,
-                    "capabilities": {"add": ["SYS_ADMIN"]},
-                },
-                "volumeMounts": volume_mounts,
-                "lifecycle": {
-                    "preStop": {
-                        "exec": {"command": ["fusermount", "-u", self.mount_point]}
-                    },
-                },
-            }
-        ]
