+++ setup.cfg
@@ -31,3 +31,8 @@ tag_svn_revision = 0
 jupyterhub.authenticators =
     egiauthenticator = egi_notebooks_hub.egiauthenticator:EGICheckinAuthenticator
     d4scienceauthenticator = egi_notebooks_hub.d4science:D4ScienceOauthenticator
+    onedataauthenticator = egi_notebooks_hub.datahub:OnedataAuthenticator
+jupyterhub.spawners =
+    egispawner = egi_notebooks_hub.egispawner:EGISpawner
+    d4sciencespawner = egi_notebooks_hub.d4science:D4ScienceSpawner
+    onedataspawner = egi_notebooks_hub.datahub:OnedataSpawner
